pyproject.toml
[tool.poetry]
name = "changie"
version = "0.1.0"
description = "Changelog generator"
authors = ["Dawid Sołtysiak <d.soltysiak51@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
coverage = {extras = ["toml"], version = "^5.1"}
pytest-cov = "^2.9.0"
pytest-mock = "^3.1.1"
flake8 = "^3.8.4"
black = "^20.8b1"
pep8-naming = "^0.11.1"

[tool.poetry.scripts]
changie = 'changie.cli:main'

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
branch = true
source = ["changie"]

[tool.coverage.report]
show_missing = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
